fix: using integers for amounts and changed getLastBlock() function to return a reference

+++ include/Transaction.hpp
@@ -11,7 +11,7 @@ struct TxIn{
 };
 
 struct TxOut{
-	double amount;
+	uint64_t amount;
 	std::string to_adress;
 };
